Moved the TypeOfList to it’s own ViewController
